[CORRECT] redo token auth, response object and error handling [TODO] update ActivityController

+++ OG-ActiTrack_server/src/main/java/com/andres_k/og/controllers/user/UserController.java
@@ -2,10 +2,13 @@ package com.andres_k.og.controllers.user;
 
 import com.andres_k.og.config.HttpResponse;
 import com.andres_k.og.config.Restricted;
+import com.andres_k.og.models.auth.ERoles;
 import com.andres_k.og.models.auth.Token;
 import com.andres_k.og.models.auth.User;
+import com.andres_k.og.services.AuthorizationService;
 import com.andres_k.og.services.TokenService;
 import com.andres_k.og.services.UserService;
+import com.andres_k.og.utils.tools.Console;
 import com.andres_k.og.utils.tools.TJson;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,41 +16,44 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
-@Restricted
+import javax.persistence.EntityNotFoundException;
+
 @Controller
 public class UserController {
     @Autowired
     private UserService userService;
     @Autowired
     private TokenService tokenService;
+    @Autowired
+    private AuthorizationService authorizationService;
 
     @RequestMapping(value = "/user/delete", method = RequestMethod.DELETE)
     @ResponseBody
     public ResponseEntity<?> deleteUser(@RequestHeader String Authorization, @RequestBody Long userId) {
-        HttpResponse response = new HttpResponse();
-
         try {
+            if (!this.authorizationService.isAuthorized(ERoles.USER, Authorization))
+                return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
             Token tokenValue = this.tokenService.getToken(Authorization);
             this.userService.deleteUser(userId, tokenValue);
-            response.addResult(true);
+            return new ResponseEntity<>(true, HttpStatus.OK);
         } catch (Exception ex) {
-            response.addError("Error deleting the user:" + ex.toString());
+            Console.log("[User/delete]: " + ex.toString());
+            return new ResponseEntity<>("Error deleting the user:" + ex.toString(), HttpStatus.NOT_FOUND);
         }
-        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
     @RequestMapping(value = "/user/update", method = RequestMethod.PUT)
     @ResponseBody
     public ResponseEntity<?> update(@RequestHeader String Authorization, @RequestBody User user) {
-        HttpResponse response = new HttpResponse();
-
         try {
+            if (!this.authorizationService.isAuthorized(ERoles.USER, Authorization))
+                return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
             Token tokenValue = this.tokenService.getToken(Authorization);
             User newUser = this.userService.updateUser(user, tokenValue);
-            response.addResult(newUser);
+            return new ResponseEntity<>(newUser, HttpStatus.OK);
         } catch (Exception ex) {
-            response.addError("Error updating the user:" + ex.toString());
+            Console.log("[User/update]: " + ex.toString());
+            return new ResponseEntity<>("Error updating the user:" + ex.toString(), HttpStatus.NOT_FOUND);
         }
-        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 }
